refactor: boilerplate module
